requirements.txt
uvicorn~=0.17.4
fastapi==0.73.0
python-multipart==0.0.5
passlib==1.7.4
PyYAML==5.4.1